+++ view/frontend/web/js/action/before-order.js
@@ -1,19 +1,23 @@
 define(
     [
+        'jquery',
         'Magento_Checkout/js/model/quote',
         'Magento_Checkout/js/model/url-builder',
         'mage/storage',
         'mage/url',
         'Magento_Checkout/js/model/error-processor',
         'Magento_Customer/js/model/customer',
-        'Magento_Checkout/js/model/full-screen-loader'
+        'Magento_Checkout/js/model/full-screen-loader',
+        'Magento_CheckoutAgreements/js/view/checkout-agreements'
     ],
-    function (quote, urlBuilder, storage, url, errorProcessor, customer, fullScreenLoader) {
+    function ($,quote, urlBuilder, storage, url, errorProcessor, customer, fullScreenLoader) {
         'use strict';
 
-        return function (paymentMethod, afterSuccess) {
+        return function (paymentMethod, agreementsCheck, afterSuccess) {
             var serviceUrl, payload;
-
+            var agreementForm = $('.payment-method._active div[data-role=checkout-agreements] input');
+            var agreementData = agreementForm.serializeArray();
+            var agreementIds = [];
 
             if (!customer.isLoggedIn()) {
                 serviceUrl = urlBuilder.createUrl('/guest-carts/:quoteId/set-payment-information', {
@@ -38,6 +42,14 @@ define(
                 };
             }
 
+            if (agreementsCheck) {
+                agreementData.forEach(function (item) {
+                    agreementIds.push(item.value);
+                });
+                payload.paymentMethod.extension_attributes = {
+                    "agreement_ids": agreementIds
+                };
+            }
             fullScreenLoader.startLoader();
 
             return storage.post(